+++ src/singleTable/adaptor/definitions/operations/crud/index.ts
@@ -2,3 +2,4 @@ export * from './delete';
 export * from './create';
 export * from './get';
 export * from './update';
+export { BaseAtomicIndexUpdate, ExpiresAtParams } from './types';