+++ src/singleTable/adaptor/definitions/operations/crud/update.ts
@@ -123,16 +123,27 @@ export class SingleTableUpdater extends BaseSingleTableOperator {
     return mixed;
   }
 
+  private ensureRemove<Entity = AnyObject>({
+    remove,
+    expiresAt,
+  }: SingleTableUpdateParams<Entity, RefConfig>) {
+    const expiresProp = expiresAt === null ? this.config.expiresAt! : null;
+
+    const actualRemove = toTruthyList([...(remove || []), expiresProp]);
+
+    if (actualRemove.length) return actualRemove as typeof remove;
+  }
+
   getUpdateParams<Entity = AnyObject>(
     params: SingleTableUpdateParams<Entity, RefConfig>,
   ): UpdateParams<Entity> {
     this.validateUpdateProps(params);
 
     const {
+      //
       conditions,
       expiresAt,
       indexes,
-      remove,
       values,
       returnUpdatedProperties,
       type,
@@ -150,8 +161,8 @@ export class SingleTableUpdater extends BaseSingleTableOperator {
       key: getPrimaryKey(params, this.config),
 
       atomicOperations: this.resolveAtomic(params),
+      remove: this.ensureRemove(params),
       conditions,
-      remove,
       returnUpdatedProperties,
 
       values: (addValues