+++ src/singleTable/model/__tests__/schema/update.test.ts
@@ -239,7 +239,7 @@ describe('single table schema - entity - update params', () => {
     });
 
     type _Tests = [
-      Expect<Equal<Params['expiresAt'], number | undefined>>,
+      Expect<Equal<Params['expiresAt'], number | undefined | null>>,
 
       Expect<Equal<Params['values'], Partial<User> | undefined>>,